# Optimize/redis n plus 1