@michaelfuest 2025-07-03
Hi @harshith-eth,

Thank you for your interest in superglue and your contribution(s).

We are currently taking a closer look at our open-source model and are in the process of setting up a detailed contribution guide. We will be closing all external open pull requests in the meantime, but will be happy to accept external contributors once we have formalized our contribution process a bit more.

Closing this for now.